refactor: 클래스 추출
